Support matching bot name when there is a unicode space separator.
Sometimes when copy and pasting messages in slack, the space character
gets pasted as \00a0, a no-break space character. The existing regex
does not recognise this as a space character and does not emit the
Response event.

This is unexpected because on slack it is indistinguishable from a
normal space character.

This commit expands the regex from \s to \p{Zs} which is the unicode
class for space characters. This is a strict superset of \s and will
catch the case where the pasted message has unicode space.